Refactor the syslog parsing code out into its own module
