+++ src/main.rs
@@ -12,9 +12,6 @@ extern crate serde;
 #[macro_use]
 extern crate serde_derive;
 extern crate serde_regex;
-#[cfg(test)]
-#[macro_use]
-extern crate serde_json;
 extern crate syslog_rfc5424;
 extern crate syslog_loose;
 
@@ -35,6 +32,7 @@ use std::collections::HashMap;
 
 mod kafka;
 mod merge;
+mod parse;
 mod rules;
 mod serve_tls;
 mod settings;
@@ -178,43 +176,6 @@ async fn accept_loop(
     Ok(())
 }
 
-#[derive(Debug)]
-enum SyslogErrors {
-    UnknownFormat,
-}
-
-#[derive(Debug)]
-struct SyslogMessage {
-    msg: String,
-}
-
-/**
- * Attempt to parse a given line either as RFC 5424 or RFC 3164
- */
-fn parse_line(line: String) -> std::result::Result<SyslogMessage, SyslogErrors> {
-    match syslog_rfc5424::parse_message(&line) {
-        Ok(msg) => {
-            return Ok(SyslogMessage {
-                msg: msg.msg,
-            })
-        },
-        Err(_) => {
-            let parsed = syslog_loose::parse_message(&line);
-
-            /*
-             * Since syslog_loose doesn't give a Result, the only way to tell if themessage wasn't
-             * parsed properly is if some fields are None'd out.
-             */
-            if parsed.timestamp != None {
-                return Ok(SyslogMessage{
-                    msg: parsed.msg.to_string(),
-                })
-            }
-            Err(SyslogErrors::UnknownFormat)
-        },
-    }
-}
-
 /**
  * connection_loop is responsible for handling incoming syslog streams connections
  *
@@ -232,7 +193,7 @@ pub async fn read_logs<R: async_std::io::Read + std::marker::Unpin>(
         let line = line?;
         debug!("log: {}", line);
 
-        let parsed = parse_line(line);
+        let parsed = parse::parse_line(line);
 
         if let Err(_e) = &parsed {
             state.metrics.counter("error.log_parse").count(1);
@@ -524,35 +485,5 @@ mod tests {
         let to_merge = r#"{"hello":"{{name}}"}"#;
         let output = perform_merge("{}", &to_merge, &state);
         assert_eq!(output, Ok("{\"hello\":\"world\"}".to_string()));
-    }
-
-    #[test]
-    fn test_parsing_invalid() {
-        let buffer = "blah".to_string();
-        let parsed = parse_line(buffer);
-        if let Ok(msg) = &parsed {
-            println!("msg: {}", msg.msg);
-        }
-        assert!(parsed.is_err());
-    }
-
-    #[test]
-    fn test_5424() {
-        let buffer = r#"<13>1 2020-04-18T15:16:09.956153-07:00 coconut tyler - - [timeQuality tzKnown="1" isSynced="1" syncAccuracy="505061"] hi"#.to_string();
-        let parsed = parse_line(buffer);
-        assert!(parsed.is_ok());
-        if let Ok(msg) = parsed {
-            assert_eq!("hi", msg.msg);
-        }
-    }
-
-    #[test]
-    fn test_3164() {
-        let buffer = r#"<190>May 13 21:45:18 coconut hotdog: hi"#.to_string();
-        let parsed = parse_line(buffer);
-        assert!(parsed.is_ok());
-        if let Ok(msg) = parsed {
-            assert_eq!("hi", msg.msg);
-        }
     }
 }
